separated main.js into modules data.js and util.js
